update build script

+++ scripts/build_release.sh
@@ -1,38 +1,48 @@
 #!/bin/sh 
 
 # config
+version=$1
 cur_pwd=`pwd`
 today=`date +%F-%T`
 
 git_uri=git@github.com:lsh123/xmlsec.git
 rpm_root=/usr/src/redhat
-build_root="$rpm_root/BUILD/xmlsec-build-area-$today"
+build_root="/tmp/xmlsec-build-area-$today"
+tar_file="xmlsec1-$version.tar.gz"
+sig_file="xmlsec1-$version.sig"
+git_version_tag=`echo $version | sed 's/\./_/g'`
 
-echo "Creating build area $build_root"
+if [ x"$version" = x ]; then
+    echo "Usage: $0 <version>"
+    exit 1
+fi
+
+echo "============== Creating build area $build_root for building xmlsec1-$version"
 rm -rf "$build_root"
 mkdir -p "$build_root"
 cd "$build_root"
 
-echo "Checking out the module '$git_url'"
+echo "============== Checking out the module '$git_url'"
 git clone $git_uri
 cd xmlsec
 find . -name ".git" | xargs rm -r
 
+echo "============== Building xmlsec1-$version"
 ./autogen.sh --prefix=/usr --sysconfdir=/etc
 make tar-release
 # can't build rpm on ubuntu
 # make rpm-release
 
-tar_file=`ls xmlsec*.tar.gz`
-echo "Moving sources tar file to $rpm_root/SOURCES/$tar_file"
-mv $tar_file $rpm_root/SOURCES
+echo "============== Signing $tar_file into $sig_file"
+gpg --output "$sig_file" --detach-sig "$tar_file"
 
-echo "Cleanup"
+echo "============== Tagging the release $version in the github"
+echo "git tag -a "xmlsec-$git_version_tag" -m 'XMLSec release $version'"
+echo "git push --follow-tags"
+
+echo "============== Move files and cleanup"
+mv "$tar_file" "$sig_file" "$cur_pwd/"
 cd "$cur_pwd"
 #rm -rf "$build_root"
-
-echo "DO NOT FORGET TO TAG THE RELEASE"
-echo "git tag -a xmlsec-1_2_N -m 'XMLSec release 1.2.N'"
-echo "git push --follow-tags"
